Completed work on the transpile preview
