+++ apps/total-typescript/src/components/code-editor/lazy-loaded-editor.tsx
@@ -10,7 +10,7 @@ export const LazyLoadedFullWidthEditor = React.lazy(() =>
 )
 
 export const LazyLoadedTranspilePreview = React.lazy(() =>
-  import('./code-editor').then((res) => {
+  import('./transpile-preview').then((res) => {
     return {
       default: res.EagerlyLoadedTranspilePreview,
     }
@@ -33,6 +33,24 @@ export const MDXEditor = (props: {children?: any}) => {
   }
 
   return <LazyLoadedFullWidthEditor code={code} language={language} />
+}
+
+export const MDXTranspilePreview = (props: {children?: any}) => {
+  // Yes, we're diving into React's internals to grab the
+  // code from the <pre> element
+  const code = props.children?.props?.children?.props?.children
+
+  if (!code) {
+    return null
+  }
+
+  let language = props.children?.props?.children?.props?.className
+
+  if (language) {
+    language = language.replace('language-', '') as Language
+  }
+
+  return <LazyLoadedTranspilePreview code={code} language={language} />
 }
 
 export const EditorTest = () => {
